Now draws sprites!

+++ include/common.h
@@ -13,10 +13,12 @@
 #define FALSE	(0)
 #define TRUE	(!FALSE)
 
-#define NELEMS(a)	(sizeof(a) / sizeof((a)[0]))
+#define NELEMS(a)	((int)(sizeof(a) / sizeof((a)[0])))
+
+#include <stdio.h>
+
 
 #ifdef DEBUG
-	#include <stdio.h>
 	#define dprintf(...)	printf(__VA_ARGS__)
 #else
 	#define dprintf(...)
